agentcore-features/code-interpreter/requirements.txt
bedrock-agentcore
strands-agents
boto3
matplotlib
numpy
